+++ src/parser/de.rs
@@ -1,6 +1,7 @@
-use std::collections::HashMap;
+use std::collections::{HashMap, LinkedList};
 
-use serde_json::Value;
+use serde::de::Error;
+use serde_json::{Map, Value};
 type PosValHolder = HashMap<String, Vec<(String, String)>>;
 
 pub fn parse_operation(path: &str) -> String {
@@ -19,7 +20,117 @@ pub fn parse_operation(path: &str) -> String {
   s
 }
 
-pub fn parse_args(p: &serde_json::Map<String, Value>) -> String {
+pub fn parse_selections_string(
+  mut defi_hm: Value,
+  input: &str,
+  operation: &str,
+) -> Result<Map<String, Value>, serde_json::Error> {
+  let mut p = String::new();
+  let mut hm = Map::new();
+  let mut curhm = &mut hm;
+  let mut queue: LinkedList<String> = LinkedList::new();
+  for char in input.chars() {
+    match char {
+      '*' => {
+        let mut tmp_defi = &mut defi_hm;
+        if let Some(first) = queue.front() {
+          if !first.eq_ignore_ascii_case(operation) {
+            queue.push_front(operation.to_owned());
+          }
+        } else {
+          queue.push_front(operation.to_owned());
+        }
+        let mut len = queue.len();
+        while len > 0 {
+          match len {
+            1 => {
+              let last = queue.pop_front().unwrap();
+              *curhm = tmp_defi
+                .get(&last)
+                .ok_or(serde_json::Error::custom(format!("404 - key: {last} not found")))?
+                .as_object()
+                .cloned()
+                .unwrap();
+            }
+            _ => {
+              let key = queue.pop_front().unwrap();
+              tmp_defi = tmp_defi
+                .get_mut(&key)
+                .ok_or(serde_json::Error::custom(format!("404 - key: {key} not found")))?;
+            }
+          }
+          len -= 1;
+        }
+        curhm = &mut hm;
+      }
+      '.' => {
+        if !p.is_empty() {
+          let pc = p.clone();
+          curhm = curhm
+            .entry(&pc)
+            .or_insert_with(|| Value::Object(Map::new()))
+            .as_object_mut()
+            .unwrap();
+          queue.push_back(pc);
+          p.clear();
+        }
+      }
+      ',' => {
+        queue.clear();
+        curhm.insert(p.clone(), Value::Null);
+        curhm = &mut hm;
+        p.clear();
+      }
+      _ => {
+        p.push(char);
+      }
+    }
+  }
+  curhm.insert(p.to_string(), Value::Null);
+  Ok(hm)
+}
+
+pub fn parse_arguments_string(matches: &str) -> Result<Map<String, Value>, serde_json::Error> {
+  let mut hm = Map::new();
+  let mut p = String::new();
+  let mut p1 = String::new();
+  let mut curhm = &mut hm;
+  let mut b = false;
+  for char in matches.chars() {
+    match char {
+      '.' => {
+        curhm = curhm
+          .entry(p.clone())
+          .or_insert_with(|| Value::Object(Map::new()))
+          .as_object_mut()
+          .unwrap();
+        p.clear();
+        b = false;
+      }
+      ',' => {
+        b = false;
+        curhm.insert(p.clone(), Value::from(p1.clone()));
+        curhm = &mut hm;
+        p.clear();
+        p1.clear();
+      }
+      '=' => {
+        b = true;
+      }
+      _ => {
+        if b {
+          p1.push(char);
+        } else {
+          p.push(char);
+        }
+      }
+    }
+  }
+  curhm.insert(p, Value::from(p1));
+  Ok(hm)
+}
+
+pub fn parse_args(p: &Map<String, Value>) -> String {
   p.iter()
     .filter(|(key, _)| !key.starts_with("api") && !key.starts_with("/api") && !key.starts_with('$'))
     .map(|(k, v)| format!("{k}={}", v.as_str().unwrap_or_default()))
@@ -32,7 +143,7 @@ pub fn parse_selections(p: &serde_json::Map<String, Value>) -> Option<String> {
 }
 
 pub fn de_kebab(qry: &str) -> String {
-  let qry = urlencoding::decode(qry).unwrap_or_default().replace('\\', "");
+  let qry = urlencoding::decode(qry).unwrap_or_default().replace(['\\', ' '], "");
   let converter = convert_case::Converter::new();
   let converter = converter.to_case(convert_case::Case::Camel);
   converter.convert(qry)
